Parser: intr. the unique_ptr<EvalContext> const_ctx_owner field
Const contexts are created only once and live for the whole lifetime
of the program, so they don't need any ownership management. But,
the address sanitizer doesn't know that and generate false positives,
claiming that the dedicated const_ctx created in ParseContext's ctor
leaks. From its point of view is right, because `const_ctx` is not
a global variable. To avoid any false positives, the `const_ctx_owner`
field has been introduced. It will *own* the const_ctx object and
it will free it, once the ParseContext is destroyed.

Note: `const_ctx` remains just a raw pointer pointing to the object owned
by `const_ctx_owner`, most of the time. That's because it's convenient for
pBlock() to temporarily assign to it a child-context, without messing with
ownership at all.